feat: support blind & attended transfers
Code is hot garbage but it does the job. The logic was changed from "To"
and "From" to simply considering a call to be a connnection of channels
an a bridge. This is because not all calls are of simply two
participants and they may change throughout its lifetime. This also
plays nicely with transfers.

Most of the transfer logic is handled by Asterisk and ARI just needs to
keep track of changes in bridge membership and update accordingly so it
can properly hangup calls when its time to.