feat: play audio for unanswered calls
If a call is allowed to go trough but it times out (far end rung but did
not pick up) or the far end outright rejects it, a brief message is now
played to explain the caller about the situation. The messge played is
to be placed in /sounds/vouncer_timeout.